Update en/1.4/getting-started/demo.md
I was cleaning my server up the other day and deleted the 
croogo.webbedit.co.uk domain forgetting why it existed, oops! Sorry!

Have restored the demo, updated CakePHP to 2.2.4 and pulled the 
latest master Croogo branch.  

Just updating the documentation as moved the demo into a subfolder 
so I can also host a demo of the 1.5 branch when it's ready. 